% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{altregress}
\alias{altregress}
\title{Select alternative method for control value regression}
\usage{
altregress(dataframe, method = "deltapheno")
}
\arguments{
\item{dataframe}{A data frame in long form that contains all of the experimental and control values.}

\item{method}{Type of regression desired. Either "deltapheno" (default), which takes each condition
data point subtracted from the mean control phenotype per strain, or "fracpheno" which takes each
condition data point divided by the mean control phenotype per strain.}
}
\value{
A data frame in long form
}
\description{
Select alternative method for control value regression
}
