% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bamfPrune.R
\name{bamfPrune}
\alias{bamfPrune}
\title{Binned Anomaly, Mitigate and Fit Pruning}
\usage{
bamfPrune(data)
}
\arguments{
\item{data}{A melted data frame to be analyzed for outliers}
}
\value{
A single data frame for a single plate file with the outlier data NA'd out.
}
\description{
Prune based on binned IQR multiples eminating from the center of the distribution. Anomalies not removd if they account for more than 5% of the population.
}

