% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{bioprune}
\alias{bioprune}
\title{Prune data based on biological cutoffs}
\usage{
bioprune(data)
}
\arguments{
\item{data}{A melted data frame to be analyzed for outliers.}

\item{drop}{A boolean stating whether observations from outlier data points
should be dropped from the returned data set. Defaults to \code{FALSE}.}
}
\value{
A data frame either with (\code{drop = TRUE}) all of the outlier data
points removed from the data frame or (\code{drop = FALSE}) three additional
columns stating whether the data point was identified as an outlier in any of
the three conditions.
}
\description{
Prune the data based on brood size and/or normalized brood size
}

