% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensureLong.R
\name{ensureLong}
\alias{ensureLong}
\title{Ensure a data frame is in long form}
\usage{
ensureLong(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame in long form
}
\description{
Ensure a data frame is in long form
}

