% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensureWide.R
\name{ensureWide}
\alias{ensureWide}
\title{Ensure a data frame is in wide form}
\usage{
ensureWide(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame in wide form
}
\description{
Ensure a data frame is in wide form
}

