% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{info}
\usage{
info(filePath, levels = 1)
}
\arguments{
\item{filePath}{The path of the file}

\item{levels}{The number of levels above the file name that identifying information is found. For "~/Dropbox/HTA/Results/20140317_GWAS1a/score/p01_control.txt" this would be 2, since 'score' is one level up and '20140317_GWAS1a' is 2 levels up.}
}
\value{
A one row data frame with a column for the date, experiment, round, assay, plate, and drug.
}
\description{
info
}

