% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvPrune.R
\name{mvPrune}
\alias{mvPrune}
\title{Remove outliers by multivariate pruning}
\usage{
mvPrune(data, mvCols = c("f.L1", "f.ad", "mean.TOF", "mean.EXT"))
}
\description{
Remove outliers by multivariate pruning
}

