% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_outliers}
\alias{prune_outliers}
\title{Prune based on outliers}
\usage{
prune_outliers(data, iqr = FALSE)
}
\arguments{
\item{data}{A melted data frame to be analyzed for outliers.}

\item{iqr}{A boolean stating whether observations should be called outlier data points
from med +/- 2*IQR. Defaults to \code{FALSE} - uses mean +/- 2*sd.}
}
\value{
A data frame either with all of the outlier data
points removed from the data frame
}
\description{
Prune based on outliers being +/- 2 IQR from the median or +/- 2 standard deviations
from the mean for each strain. Use for NIL assays with replicates, not mapping experiments.
}
