% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{readData}
\usage{
readData(fileDir, tofmin = 60, tofmax = 2000, extmin = 0,
  extmax = 10000, SVM = TRUE)
}
\arguments{
\item{fileDir}{The file or directory to be read. This can be an individual file, a directory containing .txt files to be read, or the directory containing the score and setup directories.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 60.}

\item{extmin}{The minimum extinction value allowed. Defaults to 0.}

\item{SVM}{Boolean specifying whether or not to use the support vector machine to separate worms and bubbles.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 2000.}

\item{extmin}{The maximum extinction value allowed. Defaults to 10000.}
}
\value{
If a single file is given, a single data frame for only the provided plate is returned. If an experiment directory is given, a list of two data frames will be returned. The first element in the list will be a single data frame for all of the score plates. The second element will be all of the setup plates. If there are no setup plates, NA will fill the second list element. If a setup or score directory is given, a single data frame of all of the plates in the directory will be returned.
}
\description{
Reads sorter data files into R. Built for use exlusively with worms. Depends on COPASutils for the readSorter function.
}

