% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{readfile}
\usage{
readFile(file, tofmin = 60, tofmax = 2000, extmin = 0, extmax = 10000,
  SVM = TRUE, levels = 2)
}
\arguments{
\item{file}{The file to be read.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 60.}

\item{extmin}{The minimum extinction value allowed. Defaults to 0.}

\item{SVM}{Boolean specifying whether or not to use the support vector machine to separate worms and bubbles.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 2000.}

\item{extmin}{The maximum extinction value allowed. Defaults to 10000.}
}
\value{
Returns a single data frame for a single plate file.
}
\description{
Reads sorter data files into R. Built for use exlusively with worms. Depends on COPASutils for the readSorter function.
}

