% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readPlate_worms.R
\name{readPlate_worms}
\alias{readPlate_worms}
\title{readPlate_worms}
\usage{
readPlate_worms(file, tofmin = 60, tofmax = 2000, extmin = 0,
  extmax = 10000, SVM = TRUE)
}
\arguments{
\item{file}{The file to read.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 60.}

\item{extmin}{The minimum extinction value allowed. Defaults to 0.}

\item{SVM}{Boolean specifying whether or not to use the support vector machine to separate worms and bubbles.}

\item{tofmin}{The minimum time of flight value allowed. Defaults to 2000.}

\item{extmin}{The maximum extinction value allowed. Defaults to 10000.}
}
\description{
Reads sorter data files into R. Built for use exlusively with worms. Depends on COPASutils for the readSorter function.
}

