% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readTemplate.R
\name{readTemplate}
\alias{readTemplate}
\title{Reads a template file for labeling of individual wells}
\usage{
readTemplate(templateFile, type)
}
\arguments{
\item{templateFile}{The template file to be read.}

\item{type}{The type of template being read. Can be either "strains", "conditions", "controls."}
}
\value{
Returns a data frame with 3 columns corresponing to the row, column, and value for either of the three types.
}
\description{
Reads a template file for labeling of individual wells
}

