% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{Regress control values and, optionally, assay values from experimental values}
\usage{
regress(dataframe, assay = FALSE)
}
\arguments{
\item{dataframe}{A data frame in long form that contains all of the experimental and control values.}

\item{assay}{Boolean stating whether to regress out assay as well as controls.}
}
\value{
A data frame in long form
}
\description{
Regress control values and, optionally, assay values from experimental values
}

