% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize.R
\name{remove_contamination}
\alias{remove_contamination}
\title{Remove contamination from plate data frame or list of plate data frames}
\usage{
remove_contamination(data)
}
\arguments{
\item{data}{Data object from read_data}
}
\value{
The original data object with all contaminated wells removed
}
\description{
Remove contamination from plate data frame or list of plate data frames
}

