% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizePlates.R
\name{summarizePlates}
\alias{summarizePlates}
\title{summarizePlates}
\usage{
summarizePlates(plate, strains = NULL, quantiles = FALSE, log = FALSE,
  ends = FALSE)
}
\arguments{
\item{plate}{The raw plate data frame to be summarized.}

\item{strains}{A character vector of length equal to the number of wells in in the plate being summarized.}

\item{quantiles}{Boolean indicating whether or not quantile values (q10, q25, q75, q90) should be included in the summarized data frame. Defaults to FALSE.}

\item{log}{Boolean indicating whether or not log-transformed values should be included in the summarized data frame. Defaults to FALSE.}

\item{ends}{Boolean indicating whether or not min and max values should be included in the summarized data frame. Defaults to FALSE.}
}
\description{
summarizePlates
}

