% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizeSorted.R
\name{summarizeSorted}
\alias{summarizeSorted}
\title{summarizeSorted}
\usage{
summarizeSorted(plates, strains = NULL, quantiles = FALSE, log = FALSE,
  ends = FALSE)
}
\arguments{
\item{plates}{A list consisting of all of the score plates in the first element and the setup plates in the second elements.}

\item{strains}{A character vector of length equal to the number of wells in in the plate being summarized. Defaults to NULL.}

\item{quantiles}{Boolean indicating whether or not quantile values (q10, q25, q75, q90) should be included in the summarized data frame. Defaults to FALSE.}

\item{log}{Boolean indicating whether or not log-transformed values should be included in the summarized data frame. Defaults to FALSE.}

\item{ends}{Boolean indicating whether or not min and max values should be included in the summarized data frame. Defaults to FALSE.}
}
\description{
summarizeSorted
}

