% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{sumplate}
\alias{sumplate}
\title{Summarize a plate or multiple plates}
\usage{
sumplate(plates, picked_plates = FALSE, directories = FALSE,
  quantiles = FALSE, log = FALSE, ends = FALSE, long = FALSE)
}
\arguments{
\item{plates}{A single data frame of all data or a list consisting of all of
the score plates in the first element and the setup plates in the second
elements. Alternatively, a list of data frames can be supplied (the output
from \code{read_data} if multiple directories are supplied) if the
\code{directories} flag is set to \code{TRUE}.}

\item{directories}{Set to \code{TRUE} if the supplied data comes from
mutliple directories (e.g. the RIAILs1 or RIAILs2 data). Defaults to
\code{FALSE}.}

\item{quantiles}{Boolean indicating whether or not quantile values (q10, q25,
q75, q90) should be included in the summarized data frame. Defaults to FALSE.}

\item{log}{Boolean indicating whether or not log-transformed values should be
included in the summarized data frame. Defaults to FALSE.}

\item{ends}{Boolean indicating whether or not min and max values should be
included in the summarized data frame. Defaults to FALSE.}

\item{long}{Boolean stating whether to output the data in long format.}
}
\description{
This function takes a list of plates as input and returns a single data frame
with data summarized for each well/condition pair. As input, this function
takes a two element list, with the first element being a data frame
consisting of all of the score plates to be summarized. The second element
in the list should be a data frame consisting of all of the setup plates to
be summarized, from which the number sorted to each well will be extracted.
}
