% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pruneByIQR.R
\name{bamfPrune}
\alias{bamfPrune}
\title{Removes outliers based on binned IQR}
\usage{
bamfPrune(plate)
}
\arguments{
\item{plate}{A melted data frame to be analyzed for outliers}
}
\value{
A single data frame for a single plate file with the outlier data NA'd out.
}
\description{
Removes outliers based on binned IQR
}

