% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizePlates.R
\name{summarizePlates}
\alias{summarizePlates}
\title{summarizePlates}
\usage{
summarizePlates(plate, quantiles = FALSE, log = FALSE, ends = FALSE,
  long = FALSE)
}
\arguments{
\item{plate}{The raw plate data frame to be summarized.}

\item{quantiles}{Boolean indicating whether or not quantile values (q10, q25, q75, q90) should be included in the summarized data frame. Defaults to FALSE.}

\item{log}{Boolean indicating whether or not log-transformed values should be included in the summarized data frame. Defaults to FALSE.}

\item{ends}{Boolean indicating whether or not min and max values should be included in the summarized data frame. Defaults to FALSE.}

\item{long}{Boolean indicating whether to return the data in long format. Defaults to FALSE.}

\item{strains}{A character vector of length equal to the number of wells in in the plate being summarized.}
}
\description{
This function summarizes plates when the associated setup plates are not included in the analysis. The function takes the \code{plate} argument and returns a summarized version of the plate, with one well per line. The original plate data must be in wide format.
}

