% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarizeSorted.R
\name{summarizeSorted}
\alias{summarizeSorted}
\title{Summarize a plate and a add a column for the normalized population}
\usage{
summarizeSorted(plates, strains = NULL, quantiles = FALSE, log = FALSE,
  ends = FALSE)
}
\arguments{
\item{plates}{A list consisting of all of the score plates in the first element and the setup plates in the second elements.}

\item{strains}{A character vector of length equal to the number of wells in in the plate being summarized. Defaults to NULL.}

\item{quantiles}{Boolean indicating whether or not quantile values (q10, q25, q75, q90) should be included in the summarized data frame. Defaults to FALSE.}

\item{log}{Boolean indicating whether or not log-transformed values should be included in the summarized data frame. Defaults to FALSE.}

\item{ends}{Boolean indicating whether or not min and max values should be included in the summarized data frame. Defaults to FALSE.}
}
\description{
This function takes a list of plates as input and returns a single data frame with data summarized for each well/condition pair. As input, this function takes a two element list, with the first element being a data frame consisting of all of the score plates to be summarized. The second element in the list should be a data frame consisting of all of the setup plates to be summarized, from which the number sorted to each well will be extracted.
}

